requirements.txt
google_generativeai
flask
streamlit
python-dotenv
